pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlalchemy-filterset>=2.3.0",
    "uvicorn>=0.35.0",
    "asyncpg>=0.30.0",
    "alembic>=1.16.5",
    "inject>=5.3.0",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
]
